Move new code to a dedicated package
